Back/requirements.txt

fastapi
uvicorn
python-dotenv
langchain-groq
langgraph
langchain-community
tavily-python
langchain-core